devices/pi-sender/requirements.txt
azure-eventhub==5.12.0
python-dotenv==1.0.1
board
RPi.GPIO
Adafruit-Blinka 
adafruit-circuitpython-dht
lgpio

